requirements.txt
Django==1.10.7
Fabric==1.9.0
Pillow==5.2.0
PyPDF2==1.25.1
astroid==1.4.8
coverage==3.7.1
django-easy-pdf==0.1.0
djangorestframework==3.6.3
ecdsa==0.11
futures==3.1.1
html5lib==0.9999999
isort==4.2.5
lazy-object-proxy==1.2.2
mccabe==0.5.2
mock==1.0.1
paramiko==2.4.1
pycrypto==2.6.1
pylint==1.6.4
reportlab==3.2.0
requests==2.4.3
six==1.10.0
wrapt==1.10.8
wsgiref==0.1.2
xhtml2pdf==0.0.6
yapf==0.11.0
pipdeptree==0.13.0
